Add passing tests for extrema_helper_test
